requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
nltk
transformers
python-multipart 
PyPDF2
groq